ai-agents/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
pyodbc==5.0.1
sqlalchemy==2.0.23

# AI/ML - TensorFlow (compatible versions)
tensorflow==2.15.0
keras==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# AI/ML - NLP
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99
nltk==3.8.1
spacy==3.7.2

# LangChain for LLM integration
langchain==0.0.340
langchain-community==0.0.3
langchain-openai==0.0.2

# OpenAI
openai==1.3.7

# Local LLM support
llama-cpp-python==0.2.23
huggingface-hub==0.19.4

# Vector database
chromadb==0.4.18

# Blockchain interaction
web3==6.11.3
eth-account==0.10.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
cachetools==5.3.2

